+++ policy_sentry/shared/query.py
@@ -58,3 +58,37 @@ def query_action_table(db_session, service):
         if action not in results:
             results.append(action)
     return results
+
+
+def query_action_table_by_name(db_session, service, name):
+    """Get details about an IAM Action in JSON format."""
+    rows = db_session.query(ActionTable).filter(and_(ActionTable.service.ilike(service), ActionTable.name.ilike(name)))
+    # TODO: Before submitting #29, Throw an error if it doesn't match. This is likely because the user is likely to
+    #  provide `service:action_name` as the name at first, and we want to direct them to just use the action name.
+
+    # TODO: Before submitting #29, Handle cases where the actions match multiple times - for example,
+    #  there are multiple entries for ram:TagResource due to the two different ARN formats.
+    action_table_results = {}
+    results = []
+
+    for row in rows:
+        action = row.service + ':' + row.name
+        # if row.condition_keys: # TODO: Split the condition keys based on commas.
+        # if row.dependent_actions: # TODO: Split the dependent actions based on commas.
+        # if row.dependent_actions is None:
+        #     # We have to do this otherwise the output will officially be "dependent_actions": null
+        #     dependent_actions = None
+        # else:
+        #     dependent_actions = row.dependent_actions
+        temp_dict = {
+            "action": action,
+            "description": row.description,
+            "access_level": row.access_level,
+            "resource_arn_format": row.resource_arn_format,
+            "condition_keys": row.condition_keys,
+            "dependent_actions": row.dependent_actions
+        }
+        results.append(temp_dict)
+
+    action_table_results[service] = results
+    return action_table_results